+++ app/src/main/java/com/example/wataritabi/compose/album/Album.kt
@@ -1,9 +1,9 @@
-package com.example.wataritabi.ui.compose.memo
+package com.example.wataritabi.compose.album
 
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 
 @Composable
-fun MemoScreen() {
-    Text(text = "memo")
+fun AlbumScreen() {
+    Text(text = "album")
 }
