+++ app/src/main/java/com/example/wataritabi/navigation/NavGraph.kt
@@ -4,11 +4,11 @@ import androidx.compose.runtime.Composable
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
-import com.example.wataritabi.ui.compose.album.AlbumScreen
-import com.example.wataritabi.ui.compose.destination.DestinationScreen
-import com.example.wataritabi.ui.compose.map.MapScreen
-import com.example.wataritabi.ui.compose.memo.MemoScreen
-import com.example.wataritabi.ui.compose.schedule.ScheduleScreen
+import com.example.wataritabi.compose.album.AlbumScreen
+import com.example.wataritabi.compose.destination.DestinationScreen
+import com.example.wataritabi.compose.map.MapScreen
+import com.example.wataritabi.compose.memo.MemoScreen
+import com.example.wataritabi.compose.schedule.ScheduleScreen
 
 @Composable
 fun NavGraph(navController: NavHostController) {
